requirements.txt
pillow
streamlit
huggingface_hub